Refactored flash logging into separate class Cleaned up whitespace and switched to 4-space tabs Removed unused function declaration Cleaned up values of some flash enums
