+++ Components/Flash/Driver/SPIFlash.hpp
@@ -83,8 +83,6 @@ public:
         return true;
     }
 
-
-
     /**
      * @brief Read data from flash memory at specified offset
      * @param offset offset at which to read data
