+++ Components/Flash/Driver/w25qxx.cpp
@@ -641,7 +641,6 @@ void W25qxx_WriteByte(uint8_t pBuffer, uint32_t WriteAddr_inBytes)
 	w25qxx.Lock = 0;
 }
 
-
 //###################################################################################################################
 void W25qxx_WritePage(uint8_t *pBuffer, uint32_t Page_Address, uint32_t OffsetInByte, uint32_t NumByteToWrite_up_to_PageSize)
 {
