+++ Components/Flash/Driver/w25qxx.hpp
@@ -96,7 +96,6 @@ extern "C"
 	void W25qxx_WritePage(uint8_t *pBuffer, uint32_t Page_Address, uint32_t OffsetInByte, uint32_t NumByteToWrite_up_to_PageSize);
 	void W25qxx_WriteSector(uint8_t *pBuffer, uint32_t Sector_Address, uint32_t OffsetInByte, uint32_t NumByteToWrite_up_to_SectorSize);
 	void W25qxx_WriteBlock(uint8_t *pBuffer, uint32_t Block_Address, uint32_t OffsetInByte, uint32_t NumByteToWrite_up_to_BlockSize);
-	void W25qxx_WritePageScary(uint8_t *pBuffer, uint32_t Page_Address);
 
 	void W25qxx_ReadByte(uint8_t *pBuffer, uint32_t Bytes_Address);
 	void W25qxx_ReadBytes(uint8_t *pBuffer, uint32_t ReadAddr, uint32_t NumByteToRead);
